+++ packages/core/plugin/src/withSentryAndroidGradlePlugin.ts
@@ -15,7 +15,7 @@ export interface SentryAndroidGradlePluginOptions {
   includeSourceContext?: boolean;
 }
 
-export const sentryAndroidGradlePluginVersion = '6.10.0';
+export const sentryAndroidGradlePluginVersion = '6.12.0';
 
 /**
  * Adds the Sentry Android Gradle Plugin to the project.